Line 53
Line 52 and 53 are literally the same thing. They overwrite the candidate so that  preferences[2][2] = preferences[3][2] = preferences[4][2] = 2; rather than 0. Since there is no value for preferences[2][3], [3][3], and [4][3] the output is "Invalid Vote." Rather than "3 3 1 0 " which is what it's supposed to be according to the _init_.py file. Unless I misunderstood what you are testing, this is the reason why no one has 100% on this problem, or they do and its wrong code, either way I'm not sure.