Update __init__.py
added new `check_winners` function to show help for the 3 individual print winner tests.  Had to unhide the tests in order for it to work, since the `@hidden` decorator means that the log isn't created and the help is never shown.   Having the actual help text might now be a bit redundant since the check will now show the expected vs result and it should be obvious.    (But we can't just unhide the checks as they were because it will then show an expected string that includes the regex).  
@CarterZenke if you think the help is too redundant, we can just call the `check_winner` function and if they don't match, simply return the Mismatch.