Adieu: Change stdin prompt behavior in tests
The spec requires a prompt so stdin prompt should be True.  This will also solve the long-standing issue of check50 reporting "Name: Name: ..." when the student fails to add a new line before printing the final message.
Instead they will now (correctly) see
```
:( input of "Liesl" yields "Adieu, adieu, to Liesl"
    expected "Adieu, adieu, ...", not "Name: Adieu, a..."```

Note that the prompt remains False for the EOF check, otherwise it would consume the Name: on the final line and not alert to the error there.